updated compute bounds function in physics system source file

+++ src/systems/physics_system/PhysicsSystem.cpp
@@ -253,10 +253,10 @@ namespace Project::Systems {
     const auto& circles = box->getCircles();
     for (const auto& c : circles) {
       SDL_FRect r{
-        static_cast<float>(c.x - c.r),
-        static_cast<float>(c.y - c.r),
-        static_cast<float>(c.r * Constants::CIRCLE_DIAMETER_MULTIPLIER),
-        static_cast<float>(c.r * Constants::CIRCLE_DIAMETER_MULTIPLIER)
+        c.x - c.r,
+        c.y - c.r,
+        c.r * Constants::CIRCLE_DIAMETER_MULTIPLIER,
+        c.r * Constants::CIRCLE_DIAMETER_MULTIPLIER
       };
       
       if (!hasBounds) {
